% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{ath_gene_family}
\alias{ath_gene_family}
\title{Arabidopsi: get genes by family keyword}
\usage{
ath_gene_family(keyword, ignore.case = FALSE)
}
\arguments{
\item{keyword}{character, name of the family such "ARF", "IAA", "SAUR"}

\item{ignore.case}{TRUE/FALSE, ignore case of the keyword}
}
\value{
tibble with two columns: gene_id, name
}
\description{
Helper function for extracting genes of a family (Arabidopsis).
}
\details{
Refer to \code{\link{ATHGENEANNO}} variable for full list of genes.
}
\author{
ZG Zhao
}
