% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ath_nugenes}
\alias{ath_nugenes}
\title{Arabidopsis: filter nuclear genes}
\usage{
ath_nugenes(x)
}
\arguments{
\item{x}{character vecter.}
}
\value{
character vecter.
}
\description{
Filter nuclear genes from a gene name vector (remove plastid genes).
}
\details{
This function is for Arabidopsis only.
}
\author{
ZG Zhao
}
