% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.package.R
\name{ATHGENEANNO}
\alias{ATHGENEANNO}
\alias{ATHGENGO}
\title{Arabidopsis: gene annotation data}
\description{
Both ATHGENEANNO and ATHGENGO are parsed from 'https://www.arabidopsis.org/download_files/GO_and_PO_Annotations/Gene_Ontology_Annotations/gene_association.tair.gz'.
}
\details{
Both \code{ATHGENEANNO} and \code{ATHGENGO} have \code{gene_id} column.
The \code{desc} column of \code{ATHGENEANNO} data contains a bunch of informations seperated by \code{|}. You can extract what you want.
All data are tibble objects. Run examples for more details of the object.
}
\examples{
library(vizRNAseq)
str(ATHGENEANNO)
str(ATHGENEGO)
}
\author{
ZG Zhao
}
