% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{bestBlast}
\alias{bestBlast}
\title{Get best matches from blast results.}
\usage{
bestBlast(blast_result)
}
\arguments{
\item{blast_result}{character, file path to BLAST output file in fmt6 format.}
}
\value{
tibble
}
\description{
Parsing blast results for best matches. Perl programs included: make sure you have Perl installed in your system.
}
\details{
Best results are those: (1) have longest matches in query sequence; (2) larger bit_score if matches are equal.
}
\author{
ZG Zhao
}
