% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{bestORFs}
\alias{bestORFs}
\title{Get best/longest AA sequence from fasta file (DNA strings)}
\usage{
bestORFs(fasta_file)
}
\arguments{
\item{fasta_file}{}
}
\value{
AAStringSet
}
\description{
Sequence translation
}
\details{
Sequence translation
}
\author{
ZG Zhao
}
