% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiNGO.R
\name{BiNGO_KEGG_annotation}
\alias{BiNGO_KEGG_annotation}
\title{Function BiNGO_KEGG_annotation}
\usage{
BiNGO_KEGG_annotation(df.anno, out.file, species.name)
}
\arguments{
\item{df.anno}{gene to KO annotation data.frame, whose first column is gene id and second column is KEGG ko.}

\item{out.file}{path to save the final annotation file.}

\item{species.name}{character. Name of your species.}
}
\value{
annotation file written
}
\description{
Generate customized ontology file for KEGG network analysis in Cytoscape using BiNGO app.
}
\details{
KEGG network analysis in Cytoscape using BiNGO app needs two additional files: an ontology file and an species-specific annotation file. Function "BiNGO_KEGG_annotation" is designed to generate a customized ontology file.
}
\author{
ZG Zhao
}
