% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast2go}
\alias{blast2go}
\alias{blast2kegg}
\title{Pipeline: blast2go/blast2kegg}
\usage{
blast2go(file_fmt6, file_acc_trans, file_acc2xo, uniprot = FALSE, KEGG = FALSE)
}
\arguments{
\item{file_fmt6}{character, file path to NCBI-blast output in fmt6 format}

\item{file_acc_trans}{(optional) character, file path. Data should have two columns: orignal_id and final_id. Orignal ids are the target accessions in blast results and the final ids are those for GO or KO mapping.}

\item{file_acc2xo}{character, file path. Data should have two columns: accessions and GO/KO ids.}

\item{uniprot}{TRUE/FALSE. TRUE for uniprot accessions and FALSE (default) for ncbi accessions.}

\item{KEGG}{TRUE/FALSE (default). Switch to parsing KEGG results.}

\item{max_e}{double. Default value is 1e-5.}

\item{min_score}{numeric. Default value is 1.}
}
\value{
tibble
}
\description{
Map sequences to GO or KO (KEGG) identifiers based on best blast results. Perl programs included: make sure you have Perl installed in your system.
}
\details{
Besides a blast result file (fmt6 format), you should prepare a file mapping accession to GO or KO. If the accessions in the two files are of different types, an additional mapping file is required for name translation. To prepare the mapping files, You can download the data from NCBI (for GO mapping) or KEGG (for KO mapping) website. All files input must be tab-table without header.
}
\author{
ZG Zhao
}
