% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cluster_map}
\alias{cluster_map}
\title{Simple cluster visualizations of RNASeq data from a DESeqDataSet object. Standalone or interactive through ShinyApp}
\usage{
cluster_map(typeCluster, DESeqObj, numClust)
}
\arguments{
\item{typeCluster}{either "hier" for hierarchical clustering or "km" for k-means clustering}

\item{DESeqObj}{DESeqDataSet object to use for clustering}

\item{numClust}{number of clusters for k-means cluster modeling}
}
\value{
a plot that is either a heatmap for hierarchical clustering or for k-means clustering
}
\description{
A function that uses a DESeqDataSet object to produce beautiful visualizations of clustering with the option of either
hierarchal clustering or clustering by k-means. The default setting is hierarchical. For k-means clustering, there is also
an option to change the number of clusters
}
\examples{
# To be used with the Pasilla dataset provided
cluster_map<- cluster_map(typeCluster ="km",DESeqObj = DE_convert, numClust = 4)
}
\references{
R Core Team (2020). R: A language and environment for statistical computing.
R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.

Wolfgang Huber and Alejandro Reyes (2020). pasilla: Data
package with per-exon and per-gene read counts of RNA-seq
samples of Pasilla knock-down by Brooks et al., Genome
Research 2011.. R package version 1.18.0.

Michael I. Love, S. (2020, October 27). Analyzing RNA-seq data with DESeq2.
Retrieved November 18, 2020,
from http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html

Love, M.I., Huber, W., Anders, S. Moderated estimation of
fold change and dispersion for RNA-seq data with DESeq2
Genome Biology 15(12):550 (2014)

Warner, J. (n.d.). BitBio. Retrieved December 10, 2020,
from https://2-bitbio.com/2017/10/clustering-rnaseq-data-using-k-means.html

Andrzej Oleś (2020). DEFormats: Differential gene expression data formats converter. R package version
1.18.0. https://github.com/aoles/DEFormats
}
