% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts_matrix}
\alias{cts_matrix}
\title{Matrix of read counts using the pasilla package}
\format{
A matrix with each row representing gene, and each column is a
sequenced RNA library. Values are estimated counts.
\describe{
    \item{Columns 1, 2, 3}{Treated samples}
    \item{Columns 4, 5, 6, 7}{Untreated samples}
}
}
\usage{
cts_matrix
}
\description{
This data has been extracted and is a count matrix which can be used for examples.
Data package with per-exon and per-gene read counts of RNA-seq samples of
Pasilla knock-down by Brooks et al., Genome Research 2011.
}
\references{
Wolfgang Huber and Alejandro Reyes (2020). pasilla: Data
package with per-exon and per-gene read counts of RNA-seq
samples of Pasilla knock-down by Brooks et al., Genome
Research 2011.. R package version 1.18.0.
}
\keyword{datasets}
