% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{data}
\alias{data}
\title{Useful data/variables}
\description{
constants provided in the package
}
\details{
All variables are \code{tibbles}:
\itemize{
\item ATHGENEANNO: gene annotations of Arabidopsis
\item ATHGENEGO: mappings from gene to KEGG indentifier.
\item GOFULL: data parsed from go.obo file downloaded from purl.obolibrary.org.
\item KEGGFULL: data parse from ko00001.keg file downloaded from KEGG website.
}
}
\examples{
library(vizRNAseq)
ATHGENEANNO
ATHGENEGO
GOFULL
KEGGFULL
}
\author{
ZG Zhao
}
