% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DE_convert}
\alias{DE_convert}
\title{This data is for testing purposes in preprocessing}
\format{
A DESeqDataSet object
}
\usage{
DE_convert
}
\description{
This data is for testing purposes in preprocessing
}
\keyword{datasets}
