% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{enrich_to_graph}
\alias{enrich_to_graph}
\title{network construction with enrichment data}
\usage{
enrich_to_graph(df_stat, group_label = "a")
}
\arguments{
\item{df_stat}{tibble, results from \code{\link{statKEGG}} or \code{\link{statGO}}}

\item{group_label}{character, set this if you want to join graph later with \code{\link[tidygraph:graph_join]{tidygraph::graph_join}}}
}
\value{
tbl_graph object
}
\description{
Convert enrichment results to graph/network data.
}
\details{
Refer to \code{\link[tidygraph:as_tbl_graph]{tidygraph::as_tbl_graph}} for more info about \code{tbl_graph}
}
\author{
ZG Zhao
}
