% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{getGeneFamily}
\alias{getGeneFamily}
\title{function getGeneFamily}
\usage{
getGeneFamily(gene.family, ignore.case = FALSE)
}
\arguments{
\item{gene.family}{character, name of the family such "ARF", "IAA", "SAUR"}

\item{ignore.case}{TRUE/FALSE, ignore case of the gene.family name}
}
\value{
tibble with two columns: gene_id, name
}
\description{
Helper function for extracting genes of a family (Arabidopsis).
}
\details{
Refer to \code{\link{ATHGENEANNO}} variable for full list of genes.
}
\author{
ZG Zhao
}
