% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{GO_graph}
\alias{GO_graph}
\title{Create gene ontology graph/network}
\usage{
GO_graph(obo = "full")
}
\arguments{
\item{obofile}{Character. obo file or one of 'full', 'generic', 'plant', 'pir', metagenomics'.}
}
\value{
graphNEL object
}
\description{
Gene ontology network.
}
\details{
Create graph/network from gene ontology file (obo).
}
\author{
ZG Zhao
}
