% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.package.R
\name{GOFULL}
\alias{GOFULL}
\alias{KEGGFULL}
\title{General GO and KEGG annotation data}
\description{
\code{GOFULL} is parsed from 'http://purl.obolibrary.org/obo/go.obo'. \code{KEGGFULL} is from 'https://www.genome.jp/kegg-bin/get_htext?ko00001'.
}
\details{
KEGGFULL and GOFULL are used for enrichment analysis in function \code{\link{statKEGG}} and \code{\link{statGO}}.
All data are tibble objects. Run examples for more details of the object.
}
\examples{
data(GOFULL, package='vizRNAseq')
str(GOFULL)
data(KEGGFULL, package='vizRNAseq')
str(KEGGFULL)
}
\author{
ZG Zhao
}
