% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphTerms}
\alias{graphTerms}
\title{Create shared gene network/graph.}
\usage{
graphTerms(df, percent = TRUE)
}
\arguments{
\item{df}{data.frame/tibble returned by \code{statGO} or \code{statKEGG}}

\item{percent}{TRUE/FALSE. Use percentage or number of shared genes as edge weights.}
}
\value{
list. Element 'graph': igraph object. Element 'community': igraph group object.
}
\description{
Calculate shared genes and create a comunity network.
}
\details{
Use \code{\link{sharedGenes}} for a table of shared genes only.
}
\author{
ZG Zhao
}
