% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph.R
\name{induce_KEGG_network}
\alias{induce_KEGG_network}
\title{Induce KEGG network with a given root node.}
\usage{
induce_KEGG_network(root)
}
\arguments{
\item{root}{character string, KEGG identifier from which the KEGG network will be subset.}
}
\value{
tbl_graph
}
\description{
Construct KEGG metabolic network from a given starting point.
}
\details{
Giving an KEGG identifier
}
\author{
ZG Zhao
}
