% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{KEGG_gene_map}
\alias{KEGG_gene_map}
\title{File parser: download (from KEGG website) and parse species specific KEGG pathway data}
\usage{
KEGG_gene_map(org, d.path = "KEGG/brite", force.download = FALSE)
}
\arguments{
\item{org}{character, for KEGG organism name}

\item{d.path}{character, local path having or to save KEGG brite files. Default is \code{KEGG/brite} in \code{getwd()}.}

\item{force.download}{logic, default FALSE. Force to download KEGG brite file or not.}
}
\value{
tibble with columns: gene_id, gene_name, ko and desc.
}
\description{
Get/parse gene-KO map from file
}
\details{
Download (if not existed) KEGG brite file and parse to gene-KO map.
}
\examples{
vizRNAseq::KEGG_gene_map('ath', d.path='KEGG/brite', force.download=FALSE)
}
\author{
ZG Zhao
}
