% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{KEGG_gid2pid}
\alias{KEGG_gid2pid}
\title{Function KEGG_gid2pid}
\usage{
KEGG_gid2pid(org, d.type = c("uniprot", "ncbi"), d.path = "KEGG/maps")
}
\arguments{
\item{org}{KEGG organism name (abbreviation).}

\item{d.type}{character: uniprot or ncbi}

\item{d.path}{Destination path (folder) to save or load downloaded file.}
}
\value{
data frame with two columns: gene.id and protein.id
}
\description{
Get KEGG id mapping data (gene id to uniprot or ncbi-proteinid) of a organism.
}
\details{
Retrieve id mapping data from KEGG website and save locally for later use.
}
\author{
ZG Zhao
}
