% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-KEGG.R
\name{kinfo2graph}
\alias{kinfo2graph}
\title{Function kinfo2graph}
\usage{
kinfo2graph(kos, include.genes = FALSE)
}
\arguments{
\item{kos}{character vector of KEGG indentifiers (KOs).}

\item{include.genes}{TRUE/FALSE. If true, k genes (level D entries in KEGG pathway) will be included.}
}
\value{
an igraph object
}
\description{
Construct KEGG network (graphNEL object) from a list of KOs.
}
\details{
Duplicated and cyclic edges will be removed in the final graph.
}
\author{
ZG Zhao
}
