% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mintersect}
\alias{mintersect}
\title{intersect function for list}
\usage{
mintersect(x)
}
\arguments{
\item{x}{list.}
}
\value{
vector
}
\description{
Get intersect of a list. The atoms exist in all list elements are returned.
}
\details{
The general \code{\link{intersect }} function accepts two vectors only. \code{mintersect} is more convenient for multiple intersect.
}
\author{
ZG Zhao
}
