% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{na2zero}
\alias{na2zero}
\title{Helper: convert NA to zero}
\usage{
na2zero(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
General function
}
\details{
The function name is self explained.
}
\author{
ZG Zhao
}
