% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{nugenes}
\alias{nugenes}
\title{nugenes function}
\usage{
nugenes(x)
}
\arguments{
\item{x}{character vecter.}
}
\value{
character vecter.
}
\description{
Nuclear genes
}
\details{
Get arabidopsis nuclear genes from a vecter.
}
\author{
ZG Zhao
}
