% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{onto_full_map}
\alias{onto_full_map}
\alias{KEGG_full_map}
\alias{GO_full_map}
\title{convert partial gene annotation map to full KEGG/GO tree map}
\usage{
onto_full_map(gmap, onto.type = c("ko", "go"))
}
\arguments{
\item{gmap}{data.frame with two columns: gene_id and ko (or go)}

\item{onto.type}{character scalar, accept 'ko' or 'go' only}
}
\value{
tibble with two columns: ID and genes
}
\description{
TODO
}
\details{
TODO
}
\author{
ZG Zhao
}
