% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obo.R
\name{parseOBO}
\alias{parseOBO}
\title{Parse obo file}
\usage{
parseOBO(f)
}
\arguments{
\item{f}{character, file path to obo file.}
}
\value{
tibble
}
\description{
Parse GO term info from obo format file.
}
\details{
Parse GO term info from obo format file.
}
\author{
ZG Zhao
}
