% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcharts.R
\name{plotChart}
\alias{plotChart}
\title{Function: plotChart}
\usage{
plotChart(
  dtx,
  dts,
  agi = NULL,
  alias = NULL,
  samples = NULL,
  id.label = FALSE,
  gtheme = NULL,
  sig.size = 5,
  clust = FALSE,
  del = NULL,
  num = FALSE,
  flip = FALSE,
  fname = NULL,
  width = 400,
  height = 300,
  ...
)
}
\arguments{
\item{alias}{gene name alias.}

\item{samples}{sample name alias, should be same length of exprs and pval.}

\item{gtheme}{theme object passed to ggplot2}

\item{sig.size}{integer. significant label size.}

\item{clust}{TRUE/FALSE, cluster and reorder genes.}

\item{del}{column names to be omitted from whole data.}

\item{num}{TRUE/FALSE, whether to should p values and significant labels.}

\item{flip}{TRUE/FALSE, flip xy axis.}

\item{fname}{Character: string for image (tiff) file name.}

\item{width}{figure width}

\item{height}{figure height}

\item{...}{not implemented currently.}

\item{exprs}{data.frame: rownames should be gene names and columns should be fold change data for color value.}

\item{pval}{data.frame: rownames should be gene names and columns should be p value.}

\item{genes}{character vecters for gene number subset.}
}
\description{
Plot expression diagram.
}
\details{
Plot expression diagram of seleted genes
}
\author{
ZG Zhao
}
