% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar.R
\name{plotCPM}
\alias{plotCPM}
\title{gene expression barplot}
\usage{
plotCPM(
  dfx,
  sig.size = 6,
  y.expand = 1.2,
  n.row = NULL,
  cols = NULL,
  lpos = c(0.1, 0.8),
  font_size = 16,
  x.size = 1.3
)
}
\arguments{
\item{dfx}{tibble, expression data having these columns: "gene_id", "mean", "sd", "sigLett", "treatment", "fill". Tipically generated by function \code{RSEM_pull_expression}.}

\item{sig.size}{integer. significant label size.}

\item{y.expand}{numeric, expand factor for y axis}

\item{n.row}{integer, passed to ggplot2 facet_wrap}

\item{cols}{passed to ggplot2 scale_fill_manual}

\item{lpos}{passed to ggplot2 legend.position}

\item{font_size}{passed to ggplot2 base_size}

\item{x.size}{numeric, x label size}
}
\value{
ggplot object
}
\description{
barplot using ggplot2
}
\details{
Try this: (1) get expression data with \code{RSEM_pull_expression}; (2) plot barchart with plotCPM with default params.
}
\author{
ZG Zhao
}
