% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-enrich.R
\name{plotEnrich}
\alias{plotEnrich}
\title{Visualization of enrich network}
\usage{
plotEnrich(
  egraph,
  g.layout = "igraph",
  g.algorithm = "kk",
  group.cut = 0.5,
  path.cut = 0.8,
  label.desc = TRUE,
  node.size.factor = 10,
  node.size.cex = 1,
  node.label.cex = 1,
  edge.width = 1,
  edge.arrow.angle = 30,
  edge.arrow.length = 3,
  edge.arrow.type = "open",
  x.expand = 0,
  y.expand = 0
)
}
\arguments{
\item{egraph}{tbl_graph, results from function \code{enrich_graph}.}

\item{g.layout}{param passed to \code{layout} for \code{ggraph}}

\item{g.algorithm}{layout algorithm}

\item{group.cut}{numeric (0~1), cutoff value to determine whether genes in children nodes are dominated in a parent.}

\item{path.cut}{numeric (0~1), cutoff value to determine whether genes in a node are dominated in the same pathway.}

\item{x.expand}{numeric (0~1), for \code{expand} of \code{scale_x_continuous}.}

\item{y.expand}{numeric (0~1), for \code{expand} of \code{scale_x_continuous}.}
}
\value{
ggraph
}
\description{
Plot network of created from KEGG/GO enrichment results
}
\details{
Joined tbl_graphs are support. Please pipe \code{facet_nodes} after \code{plotEnrich}.
}
\author{
ZG Zhao
}
