% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar.R
\name{plotFPKM}
\alias{plotFPKM}
\title{plotFPKM function.}
\usage{
plotFPKM(
  dtx,
  agi = NULL,
  alias = NULL,
  gids = NULL,
  filename = NULL,
  width = 400,
  height = 300,
  sig.size = 12,
  y.expand = 1.05,
  nrow = NULL,
  ncol = NULL,
  cols = c("gray80", "gray30"),
  gtheme = NULL
)
}
\arguments{
\item{dtx}{data.frame}

\item{agi}{agi for data subset}

\item{alias}{gene name alias}

\item{gids}{gene_id filters.}

\item{filename}{image filename}

\item{width}{figure width}

\item{height}{figure height}

\item{sig.size}{size of significant labels}

\item{y.expand}{expand factor of y upper limit}

\item{nrow}{}

\item{ncol}{}

\item{cols}{fill colors of bars}

\item{gtheme}{additional theme set for ggplot object.}
}
\description{
plotFPKM barchart
}
\details{
Deprecated. Use plotCPM instead.
}
\author{
ZG Zhao
}
