% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heat.R
\name{plotGHeat}
\alias{plotGHeat}
\title{Deprecated: plotGHeat}
\usage{
plotGHeat(
  dtx,
  gids,
  treats = NULL,
  cut.k = 8,
  v.cut = 0.5,
  border = "gray",
  del = NULL,
  flip = FALSE,
  gtheme = NULL,
  fill.mid = "white",
  fill.low = "steelblue",
  fill.high = "sienna4",
  fill.legend = TRUE
)
}
\arguments{
\item{dtx}{data.frame}

\item{gids}{names of gene subset for plotting.}

\item{treats}{Sample name labels.}

\item{cut.k}{groups to cut (for cutree).}

\item{border}{color for geom_tile.}

\item{del}{columns to omit.}

\item{flip}{TRUE/FALSE. flip xy}

\item{gtheme}{additional ggplot2 theme sets.}

\item{fill.mid}{mid color for heatmap.}

\item{exprs}{Expression or log2 fold change data. First column should be gene_id.}
}
\description{
Deprecated. Use \code{vheatmap} please.
}
\details{
Select a subset of genes and plot expression heatmap.
}
\author{
ZG Zhao
}
