% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.maps.R
\name{plotHeat}
\alias{plotHeat}
\title{Function: plotHeat}
\usage{
plotHeat(
  datax,
  gids = NULL,
  agi = NULL,
  alias = NULL,
  fname = NULL,
  width = 800,
  height = 200,
  v.cut = NULL,
  fontsize = 9,
  fontsize_row = 12,
  fontsize_col = 6,
  cut.k = 20,
  border = NA,
  clust.row = FALSE,
  ...
)
}
\arguments{
\item{datax}{data.frame. Columns "gene_id" and "agi" are required together with value columns.}

\item{gids}{gene_id to subset.}

\item{agi}{agi to subset.}

\item{alias}{Gene name alias.}

\item{fname}{File path for save tiff figure.}

\item{width}{Figure width.}

\item{height}{Figure height}

\item{fontsize}{Base font size}

\item{fontsize_row}{Font size for row labels.}

\item{fontsize_col}{Font size for column labels.}

\item{cut.k}{Number to cut.}

\item{clust.row}{TRUE/FALSE.}

\item{...}{Further params passed to pheatmap function.}
}
\description{
Plot heatmap with pheatmap
}
\details{
Heamap plot function.
}
\author{
ZG Zhao
}
