% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.maps.R
\name{plotHeatAP}
\alias{plotHeatAP}
\title{plotHeatAP}
\usage{
plotHeatAP(
  dtx,
  agi = NULL,
  gids = NULL,
  apclust = TRUE,
  fname = NULL,
  width = 400,
  height = 600,
  v.cut = 6,
  ...
)
}
\arguments{
\item{dtx}{}

\item{agi}{}

\item{gids}{}

\item{fname}{}

\item{width}{}

\item{height}{}

\item{v.cut}{}

\item{...}{}
}
\description{
plotHeatAP
}
\details{
plotHeatAP
}
\author{
ZG Zhao
}
