% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-KEGG.R
\name{plotKEGG}
\alias{plotKEGG}
\title{Plot KEGG network or igraph object.}
\usage{
plotKEGG(
  info,
  layout_method = "fdp",
  p.cut = 0.05,
  plot.genes = FALSE,
  edge.col = "black",
  cex.label = 1,
  cex.node = 1,
  cex.arrowsize = 1,
  cex.bar.label = 1,
  bar.adj = 0,
  label.description = TRUE,
  ...
)
}
\arguments{
\item{info}{accept two type of data: an igraph object, or KEGG.stats object obtained by \code{\link{KEGG_enrich}}.}

\item{layout_method}{The layout method to use: One of dot, neato, twopi, circo, and fdp. The default is dot}

\item{p.cut}{cut off value of BH adjust p value. Node with greater p value may not be included in the graph if it is not a parent node of any selected nodes.}

\item{plot.genes}{TRUE or FALSE (default). If TRUE, k genes will be shown in the graph.}

\item{edge.col}{edge.color}

\item{cex.label}{number, scaling factor for node label.}

\item{cex.node}{number, scaling factor for node size (shape).}

\item{cex.arrowsize}{number, scaling factor for arrow size.}

\item{cex.bar.label}{number, scaling factor for color bar indicators (p values).}

\item{bar.adj}{integer. Adjustment (horizontal) of color bar.}

\item{label.description}{TRUE/FALSE. If true, replace node label with pathway description.}

\item{...}{refer to \code{\link{Rgraphviz::GraphvizAttributes}}}
}
\description{
Plot KEGG networks, mostly for results obtained by \code{\link{KEGG_enrich}}.
}
\details{
Refer to \code{\link{graph::plot.graphNEL}} for more details.
}
\author{
ZG Zhao
}
