% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vennPlot.R
\name{plotVenn}
\alias{plotVenn}
\title{plotVenn function}
\usage{
plotVenn(
  dtx,
  fname,
  cnames = NULL,
  fig.size = 600,
  adj.cat = 0,
  adj.pos = 0,
  adj.mar = 0,
  cex.num = 0,
  cex.cat = 0
)
}
\arguments{
\item{dtx}{list of length 3 or 4.}

\item{fname}{String, file name to save.}

\item{cnames}{String vector for category names.}

\item{fig.size}{Numeric. Figure size.}

\item{adj.cat}{}

\item{adj.pos}{}

\item{adj.mar}{}

\item{cex.num}{}

\item{cex.cat}{}
}
\description{
plot venn diagram.
}
\details{
Plot venn diagram with list data of length 3 or 4.
}
\author{
ZG Zhao
}
