% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{read_DEGs}
\alias{read_DEGs}
\title{function read_DEGs}
\usage{
read_DEGs(files, g.union = FALSE)
}
\arguments{
\item{files}{file paths to "DEG" files generated by \code{\link{RSEM_pipeline}}}

\item{g.union}{TRUE/FALSE. Get union or intersect of gene ids.}
}
\value{
character vector of gene ids
}
\description{
Read and combine (intersect or union) gene ids.
}
\details{
Convenient function to perform intersection of union.
}
\author{
ZG Zhao
}
