% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cuff.R
\name{readFPKM}
\alias{readFPKM}
\title{Function readFPKM}
\usage{
readFPKM(
  xfiles,
  treats,
  tr.levels = NULL,
  std = TRUE,
  sig = TRUE,
  xpkm = FALSE,
  control = ""
)
}
\arguments{
\item{xfiles}{filename}

\item{treats}{Character vector of same length of filenames for names of treatments.}

\item{tr.levels}{Character vector of 2 for treatment levels: control and treatment.}

\item{std}{TRUE/FALSE. combine stdev or not}

\item{sig}{TRUE/FALSE. combine significance label or not.}

\item{xpkm}{return normalized FPKM values}

\item{control}{If all samples share a common control, provide its name.}
}
\value{
data.frame
}
\description{
Function for Cuff: read FPKM data
}
\details{
Read FPKM data from files, with or without stdev and significant tr.levels.
}
\author{
ZG Zhao
}
