% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.melt.R
\name{readGIDs}
\alias{readGIDs}
\title{Function readGIDs}
\usage{
readGIDs(files, reg = c("all", "up", "down"), agi = FALSE, fc = 1.5,
  labels = NULL, FPKM.cut = 0, q.cut = 1, p.cut = 1, ensemble = NULL)
}
\arguments{
\item{files}{expression file}

\item{reg}{regulation type: all, up or down}

\item{agi}{TRUE/FALSE. Return agi or gene_id instead.}

\item{fc}{fold change filter}

\item{labels}{String vector for data labels.}

\item{FPKM.cut}{Numeric. FPKM cut value.}

\item{q.cut}{Numeric. q_value cut threshold.}

\item{ensemble}{String vector.}

\item{cc}{column to return}
}
\description{
readGIDs
}
\details{
readGIDs read in gene_id or agi
}
\author{
ZG Zhao
}
