% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{reg_norm_res}
\alias{reg_norm_res}
\title{Runs DESeq2 pipeline and returns both log2 and variance stabilized log2 transformed data}
\usage{
reg_norm_res(DE_convert)
}
\arguments{
\item{DE_convert}{A DESeqDataSet object}
}
\value{
A vector containing both log2 and variance stabilized log2 transformed data as DESeqDataSet objects.
}
\description{
A function that takes in a DESeqDataSet and builds a result table which can be used
to perform cluster analysis. This function is helpful if already using Bioconductor workflow
 and already have a DESeqDataSet object.
}
