% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.voom.R
\name{RSEM.comparison}
\alias{RSEM.comparison}
\title{RSEM data preparation: logFC and p values}
\usage{
RSEM.comparison(
  d.path,
  comps.treatment,
  cmps.control,
  o.path = NULL,
  g.type = "gene",
  d.type = "fc"
)
}
\arguments{
\item{d.path}{path to the folder containing "gene.count.matrix" files}

\item{comps.treatment}{sample names (treatments)}

\item{cmps.control}{sample names (controls)}

\item{o.path}{output file name.}

\item{g.type}{gene type: "gene" or "isoform"}

\item{d.type}{data type: "fc" for logFC and "p" for PValue column}
}
\value{
data.frame whose first column is "gene_id"
}
\description{
For RSEM-voom data: Get comparison data of with given treatment and control sample names.
}
\details{
Raw data is generated by "run_DE_analysis.pl --method voom" (trinityrnaseq unitility)
}
\author{
zhao
}
