% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.voom.R
\name{RSEM.gene_counts}
\alias{RSEM.gene_counts}
\title{RSEM data preparation: gene count data}
\usage{
RSEM.gene_counts(d.path, o.path = NULL, g.type = "gene", d.type = "raw")
}
\arguments{
\item{d.path}{path to the folder containing "gene.count.matrix" files}

\item{o.path}{output file name.}

\item{g.type}{gene type: "gene" or "isoform"}

\item{d.type}{data type: "raw", "mean" or "sd". Out file of "raw" data can be applied directly in "run_DE_analysis.pl --method voom" downstream analysis.}
}
\value{
matrix
}
\description{
For RSEM data: read in and merge gene count data into a single matrix.
}
\details{
Gene expressions are estimated by "abundance_estimates_to_matrix.pl" bundled with trinityrnaseq software using "--est_method RSEM".
}
\author{
zhao
}
