% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{RSEM_comparison}
\alias{RSEM_comparison}
\title{Two-sample comparison for DEG identification.}
\usage{
RSEM_comparison(d.path, control, treatment, g.type = "gene")
}
\arguments{
\item{d.path}{path to the folder containing "gene.count.matrix" files}

\item{control}{character, control sample (single)}

\item{treatment}{character, treatment sample (single)}

\item{g.type}{gene type: "gene" or "isoform"}
}
\value{
tibble
}
\description{
RSEM only function
}
\details{
Refer to function \code{DEGlist} in \code{edgeR} and \code{lmFit} in \code{limma} package.
}
\author{
ZG Zhao
}
