% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{RSEM_pipeline}
\alias{RSEM_pipeline}
\title{Data export after RSEM abundance estimated.}
\usage{
RSEM_pipeline(
  data_dir,
  treatments,
  controls,
  output_dir = "data",
  isoform = FALSE,
  DEG.fc = 1,
  annotations = NULL
)
}
\arguments{
\item{data_dir}{file path, typically the RSEM output folder.}

\item{treatments}{character vector, sample names.}

\item{controls}{character vector, sample names.}

\item{output_dir}{file path}

\item{isoform}{TRUE/FALSE (default). FALSE for genes, TRUE for isoforms.}

\item{DEG.fc}{numeric, log(fold_changed) threshold.}

\item{annotations}{supply a gene annotation data.frame with \code{gene_id} column if you want to include gene annotation for DEGs.}
}
\value{
NULL. Data will be saved to output dir.
}
\description{
Before running this pipeline, you should get expression matrices using Trinity command: \verb{abundance_estimates_to_matrix.pl --est_method RSEM}.
}
\details{
Pipeline: export expression, run comparisons and get DEG lists.
}
\author{
ZG Zhao
}
