% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{RSEM_pull_expression}
\alias{RSEM_pull_expression}
\title{RSEM only: pull/extract after \code{\link{RSEM_pipeline}}}
\usage{
RSEM_pull_expression(data_dir, gids, smps)
}
\arguments{
\item{data_dir}{character, directory/folder where RSEM pipeline results saved.}

\item{gids}{character vector, gene ids.}

\item{smps}{character vector, sample names.}
}
\value{
tibble
}
\description{
Get expression data (mean, sd and significant letters) for visulization (plotCPM )
}
\details{
Get expression data
}
\author{
ZG Zhao
}
