% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_info}
\alias{sample_info}
\title{Matrix of read counts using the pasilla package}
\format{
A matrix with each row representing gene, and each column is a
sequenced RNA library. Values are estimated counts.
\describe{
    \item{Columns 1}{Condition}
    \item{Columns 2}{Type}
}
}
\usage{
sample_info
}
\description{
This data has been extracted and is contains sample information which can be used in
conjunction with the provided count matrix for examples.
Data package with per-exon and per-gene read counts of RNA-seq samples of
Pasilla knock-down by Brooks et al., Genome Research 2011.
}
\keyword{datasets}
