% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{sharedGenes}
\alias{sharedGenes}
\title{Calculate shared genes.}
\usage{
sharedGenes(df, percent = FALSE)
}
\arguments{
\item{df}{data.frame/tibble returned by \code{statGO} or \code{statKEGG}}

\item{percent}{TRUE/FALSE. If TRUE, return percentage of genes (union of two nodes).}
}
\value{
tibble with three columns: v1, v2 and w (number or percent of shared genes).
}
\description{
Calculate shared genes of each pair of GO terms or KEGG pathways.
}
\details{
Calculate shared genes of each pair of GO terms or KEGG pathways.
}
\author{
ZG Zhao
}
