% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bar.R
\name{theme_sci}
\alias{theme_sci}
\title{theme_sci ggplot2 theme.}
\usage{
theme_sci(base_size = 12, base_family = "Helvetica", ...)
}
\arguments{
\item{...}{args passed to theme_bw().}
}
\value{
ggplot2 theme set
}
\description{
theme_sci
}
\details{
ggplot2 theme_sci, a clear theme for figures of scientific papers.
}
\author{
ZG Zhao
}
