% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{uniprot2kgene}
\alias{uniprot2kgene}
\title{Function uniprot2kgene}
\usage{
uniprot2kgene(uids, db.files = NULL)
}
\arguments{
\item{uids}{chr vector}

\item{db.files}{additional database (uniprot to k gene mapping) file.}
}
\value{
data.frame with two columns: uniprot and kgene
}
\description{
Map uniprot ids to KEGG (K gene)
}
\details{
Map uniprot ids to K genes. Databases of there species are included: ath, osa and hsa. More databases (files) can be included by db.files whose first two columns should be uniprot id and kgene.
}
\author{
ZG Zhao
}
