% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{updateDB_u2k}
\alias{updateDB_u2k}
\title{Function updateDB_u2k}
\usage{
updateDB_u2k(db.file, uids, overwirte = FALSE)
}
\arguments{
\item{db.file}{}

\item{uids}{chr vector}

\item{overwirte}{TRUE/FALSE, write back to db.file if TRUE.}
}
\value{
data.frame
}
\description{
Update the uniprot to kgene mapping file.
}
\details{
Update the uniprot to kgene mapping file.
}
\author{
ZG Zhao
}
