% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-heat.R
\name{vheatmap}
\alias{vheatmap}
\title{plot heatmap}
\usage{
vheatmap(mat, vcut = 4, border = TRUE, ...)
}
\arguments{
\item{mat}{matrix or data.frame/tibble. Refer to \code{\link{pheatmap::pheatmap}}}

\item{vcut}{numeric. Cut (max) value.}

\item{border}{TRUE/FALSE.}

\item{...}{params passing to pheatmap}
}
\description{
Improved version of pheatmap
}
\details{
Improvements: (1) white color is fixed to zero; (2) use okabe colors.
}
\author{
ZG Zhao
}
