% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{bestMatches}
\alias{bestMatches}
\title{Get best matches from blast results.}
\usage{
bestMatches(blast_result, blast_db, prot = FALSE)
}
\arguments{
\item{blast_result}{BLAST out put file in fmt6 format}

\item{blast_db}{BLAST db file (fasta) when you do blast}

\item{prot}{TRUE/FALSE. Protein blast db (not query sequences).}
}
\value{
tibble
}
\description{
parse blast result
}
\details{
parse blast result.
}
\author{
ZG Zhao
}
