% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiNGO.R
\name{BiNGO_KEGG_ontology}
\alias{BiNGO_KEGG_ontology}
\title{Data convert: generate KEGG ontology file for BiNGO (Cytoscape plugin)}
\usage{
BiNGO_KEGG_ontology(out.file)
}
\arguments{
\item{out.file}{file name for output}
}
\value{
out.file written
}
\description{
Generate customized ontology file for KEGG network analysis in Cytoscape using BiNGO app.
}
\details{
KEGG network analysis in Cytoscape using BiNGO app needs two additional files: a general ontology file and an species-specific annotation file. This function creates the former.
}
\author{
ZG Zhao
}
