% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{mat_conversion}
\alias{mat_conversion}
\title{Converts a count matrix and table of sample information to a DESeqDataSet object.}
\usage{
mat_conversion(RNA_count_data, sample_info_data)
}
\arguments{
\item{RNA_count_data}{A matrix with each row representing gene, and each column is a sequenced RNA library.Values are estimated counts.}

\item{sample_info_data}{A table of sample information that has metadata on count table columns}
}
\value{
A DESeqDataSet object for cluster analysis
}
\description{
A function that uses the DESeq pipeline to convert a count matrix and table of sample.
This is for preprocessing the data to get it ready for cluster analysis. Additionally, as of present release,
this function performs conversion of matrix with multiple variables- ~ group + condition
}
\examples{

 # Example 1:
 # Using pasilla dataset available with package

 mat_conversion <- mat_conversion(cts_matrix, sample_info)

}
\references{
Wolfgang Huber and Alejandro Reyes (2020). pasilla: Data
package with per-exon and per-gene read counts of RNA-seq
samples of Pasilla knock-down by Brooks et al., Genome
Research 2011.. R package version 1.18.0.

Michael I. Love, S. (2020, October 27). Analyzing RNA-seq data with DESeq2.
Retrieved November 18, 2020,
from http://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html

Love, M.I., Huber, W., Anders, S. Moderated estimation of
fold change and dispersion for RNA-seq data with DESeq2
Genome Biology 15(12):550 (2014)
}
