% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mintersect}
\alias{mintersect}
\title{Multiple intersect function}
\usage{
mintersect(x)
}
\arguments{
\item{x}{list.}
}
\value{
vector
}
\description{
Multiple intersect
}
\details{
Get intersect of a list. The atoms exist in all list elements are returned.
}
\author{
ZG Zhao
}
