% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plotHeat}
\alias{plotHeat}
\title{Function: plotHeat}
\usage{
plotHeat(
  datax,
  gids = NULL,
  agi = NULL,
  alias = NULL,
  fname = NULL,
  trans.axis = FALSE,
  v.cut = 4,
  cut.k = 20,
  border = NA,
  clust.row = FALSE,
  clust.col = TRUE,
  ...
)
}
\arguments{
\item{datax}{data.frame. Columns "gene_id" and "agi" are required together with value columns.}

\item{gids}{gene_id to subset.}

\item{agi}{agi to subset.}

\item{alias}{Gene name alias.}

\item{cut.k}{Number to cut.}

\item{clust.row}{TRUE/FALSE.}

\item{...}{Further params passed to pheatmap function.}

\item{width}{Figure width.}

\item{height}{Figure height}
}
\description{
Plot heatmap with pheatmap
}
\details{
Heamap plot function.
}
\author{
ZG Zhao
}
