% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.RSEM.R
\name{RSEM_expression}
\alias{RSEM_expression}
\title{RSEM results: calculate CPM}
\usage{
RSEM_expression(d.path, g.type = "gene", d.type = "mean")
}
\arguments{
\item{d.path}{path to the folder containing "gene.count.matrix" files}

\item{g.type}{gene type: "gene" or "isoform"}

\item{d.type}{data type: "mean" or "sd"}
}
\value{
tibble, first column is "gene_id"
}
\description{
For RSEM data: calculate CPM (counts per million aligned reads)
}
\details{
Raw data is generated by "abundance_estimates_to_matrix.pl --est_method RSEM" (trinityrnaseq unitility)
}
\author{
zhao
}
