% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statKEGG}
\alias{statKEGG}
\title{KEGG pathway enrichment analysis.}
\usage{
statKEGG(gids, gmap, p.cut = 0.05, refGenes = NULL)
}
\arguments{
\item{gids}{character vector: gene ids to analysis}

\item{gmap}{data.frame or tibble with gene_id and ko (KEGG identifiers for genes such as K00185).}

\item{p.cut}{cut off value of BH adjust p value. Hypergeometric test is applied.}

\item{refGenes}{character vector of reference gene set.}
}
\value{
tibble
}
\description{
The analysis requires two data sets: a gene list and a gene-to-ko mapper. A gene-to-ko list can be retrieved from KEGG website or obtained by sequence blast.
}
\details{
KEGG pathway enrichment analysis.
}
\author{
ZG Zhao
}
